lib/mito/io/vtk: move {vtk} writers to {vtk} directories
