+++ lib/mito/io/public.h
@@ -10,24 +10,14 @@
 // external packages
 #include "externals.h"
 
-// get the forward declarations
-#include "forward.h"
-
-// published type factories; this is the file you are looking for...
-#include "api.h"
+// classes implementation
+#include "GridWriter.h"
 
 // classes implementation
 #include "summit/public.h"
 #ifdef WITH_VTK
 #include "vtk/public.h"
 #endif    // WITH_VTK
-
-// classes implementation
-#include "MeshWriter.h"
-#include "PointCloudWriter.h"
-
-// factories implementation
-#include "factories.h"
 
 
 // end of file
