+++ lib/mito/io/vtk/public.h
@@ -10,8 +10,23 @@
 // external packages
 #include "externals.h"
 
+// get the forward declarations
+#include "forward.h"
+
+// published type factories; this is the file you are looking for...
+#include "api.h"
+
 // wrappers
 #include "vtk_cell.h"
 #include "vtk_point.h"
+
+// classes
+#include "GridWriterVTK.h"
+#include "MeshWriterVTK.h"
+#include "PointCloudWriterVTK.h"
+
+// factories implementation
+#include "factories.h"
+
 
 // end of file
